10 10/2

+++ src/BlogExpert.Negocio/Services/AutorService.cs
@@ -1,17 +1,20 @@
 ﻿using BlogExpert.Negocio.Entities;
 using BlogExpert.Negocio.Entities.Validations;
 using BlogExpert.Negocio.Interfaces;
+using BlogExpert.Negocio.Seguranca;
 
 namespace BlogExpert.Negocio.Services
 {
     public class AutorService : ServiceBase, IAutorService
     {
         private readonly IAutorRepository _autorRepository;
+
         public AutorService(IAutorRepository autorRepository, INotificador notificador) : base(notificador)
         {
             _autorRepository = autorRepository;
         }
-        public async Task Adicionar(Autor autor)
+
+        public async Task Adicionar(Autor autor, ContaAutenticada contaAutenticada)
         {
             if (!ExecutarValidacao(new AutorValidation(), autor)) return;
 
@@ -21,10 +24,12 @@ namespace BlogExpert.Negocio.Services
                 return;
             }
 
+            autor.EmailCriacao = contaAutenticada.EmailConta;
+
             await _autorRepository.Adicionar(autor);
         }
 
-        public async Task Atualizar(Autor autor)
+        public async Task Atualizar(Autor autor, ContaAutenticada contaAutenticada)
         {
             if (!ExecutarValidacao(new AutorValidation(), autor)) return;
 
@@ -34,10 +39,12 @@ namespace BlogExpert.Negocio.Services
                 return;
             }
 
+            if (!VerificarSePodeManipularAutor(autor, contaAutenticada)) return;
+
             await _autorRepository.Atualizar(autor);
         }
 
-        public async Task Remover(Guid id)
+        public async Task Remover(Guid id, ContaAutenticada contaAutenticada)
         {
             var autor = await _autorRepository.ObterPorId(id);
 
@@ -53,12 +60,22 @@ namespace BlogExpert.Negocio.Services
                 return;
             }
 
+            if (!VerificarSePodeManipularAutor(autor, contaAutenticada)) return;
+
             await _autorRepository.Remover(id);
         }
 
         public void Dispose()
         {
             _autorRepository?.Dispose();
+        }
+
+        private bool VerificarSePodeManipularAutor(Autor autor, ContaAutenticada contaAutenticada)
+        {
+            if (contaAutenticada.EhAdministrador || autor.EmailCriacao == contaAutenticada.EmailConta || autor.Email == contaAutenticada.EmailConta) return true;
+            
+            Notificar("A conta autenticada não pode manipular esse autor.");
+            return false;
         }
     }
 }
